Move each auth client to own file.
