+++ packages/auth/src/authClient/netlify.ts
@@ -1,0 +1,43 @@
+import type NetlifyIdentityNS from 'netlify-identity-widget'
+
+import type { AuthClient } from './'
+
+export type NetlifyIdentity = typeof NetlifyIdentityNS
+
+export const mapAuthClientNetlify = (client: NetlifyIdentity): AuthClient => {
+  return {
+    type: 'netlify',
+    client,
+    login: () => {
+      return new Promise((resolve, reject) => {
+        let autoClosedModal = false
+        client.open('login')
+        client.on('login', (user) => {
+          // This closes the modal which pops-up immediately after you login.
+          autoClosedModal = true
+          client.close()
+          return resolve(user)
+        })
+        client.on('close', () => {
+          !autoClosedModal && resolve(null)
+        })
+        client.on('error', reject)
+      })
+    },
+    logout: () => {
+      return new Promise((resolve, reject) => {
+        client.logout()
+        client.on('logout', resolve)
+        client.on('error', reject)
+      })
+    },
+
+    getToken: async () => {
+      const user = await client.currentUser()
+      return user?.token?.access_token || null
+    },
+    currentUser: async () => {
+      return client.currentUser()
+    },
+  }
+}
